+++ test/import-service-test.js
@@ -7,6 +7,7 @@ var path = require('path')
 var ImportService = require('../lib/import-service.js')
 var Logger = require('koop-logger')
 var log = new Logger({})
+var Files = require('koop/lib').Files
 
 var importService
 var fakeCache = {
@@ -26,18 +27,7 @@ describe('importing a feature service into the cache', function () {
       log: log,
       cache: fakeCache,
       itemTitle: 'foobar',
-      files: {
-        createWriteStream: function () {
-          return {
-            write: function () {},
-            end: function () {},
-            on: function () {},
-            abort: function () {},
-            emit: function () {},
-            removeListener: function () {}
-          }
-        }
-      }
+      files: new Files({config: {data_dir: './test/data'}})
     })
 
     // speed up the backoff so the test runs faster and does not time out